requirements.txt
discord.py
emoji~=2.9.0